% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrm.R
\name{all2curv}
\alias{all2curv}
\title{\code{all2curv} is a helper function that takes in a data frame consisting of ST profile and its fragment ID
and combine them with the fragment sequences and fragment curve number}
\usage{
all2curv(st_loci, alleles)
}
\arguments{
\item{st_loci}{is data frame of ST_profile and its corresponding fragment ID}

\item{alleles}{is the data frame generated using \code{alleles}}
}
\value{
data frame with ST, fragments sequences, fragments curve numbers
}
\description{
\code{all2curv} is a helper function that takes in a data frame consisting of ST profile and its fragment ID
and combine them with the fragment sequences and fragment curve number
}
