% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{branch.percent}
\alias{branch.percent}
\title{\code{branch.percent} is used to calculate 1 or more percentage of dissimilarity and list the position with highest percentage.}
\usage{
branch.percent(seq, targets, level = 1, included = NULL,
  excluded = NULL, numRes = 1)
}
\arguments{
\item{seq}{is fastaDNA object to analysed.}

\item{targets}{is the name of the reference allele.}

\item{level}{is the number of positions.}

\item{included}{is the included position for analysed, these will force the computation
to compute the percentage at the position no matter what}

\item{excluded}{is the positions that are excluded from computation}

\item{numRes}{is the number of result to be returned.}
}
\value{
Will returns percentage of dissimilarity and the position
as well as the residual positions (i.e. those positions with 100% on their own).
}
\description{
\code{branch.percent} is used to calculate 1 or more percentage of dissimilarity and list the position with highest percentage.
}
