% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrm.R
\name{calGC}
\alias{calGC}
\title{\code{calGC} is used to calculate the number of GC content in the sequences}
\usage{
calGC(seq)
}
\arguments{
\item{seq}{is the sequence to be calculated}
}
\value{
will returns the sum of GC content in the sequence
}
\description{
\code{calGC} is used to calculate the number of GC content in the sequences
}
