% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrm.R
\name{changeCurve}
\alias{changeCurve}
\title{\code{changeCurve} is used to modify the curve number if the conditions are met. To specify conditions, create a csv file with the following headings:
fragment_name, type, replace_with, if_match_fragment, if_match_curve,
where type is
1. replace if match fragment
2. replace if match curve number
for non-relevant field, use _}
\usage{
changeCurve(changeSet = NULL, c_st_all)
}
\arguments{
\item{changeSet}{is the location of the csv created}

\item{c_st_all}{is the dataframe generated using \code{combineLoci}}
}
\value{
the dataframe with curve number modified. All rows with NAs are also dropped. All columns with fragment sequence also dropped.
}
\description{
\code{changeCurve} is used to modify the curve number if the conditions are met. To specify conditions, create a csv file with the following headings:
fragment_name, type, replace_with, if_match_fragment, if_match_curve,
where type is
1. replace if match fragment
2. replace if match curve number
for non-relevant field, use _
}
