% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{Chlamydia_mapped}
\alias{Chlamydia_mapped}
\title{Normal Fasta sample file
Chlamydia sequence types with no deletion}
\description{
Normal Fasta sample file
Chlamydia sequence types with no deletion
}
