% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrm.R
\name{combineLoci}
\alias{combineLoci}
\title{\code{combineLoci} is used to organise the fragment to its respective ST}
\usage{
combineLoci(st_profile, alleles, changeSet = NULL, included = NULL,
  excluded = NULL)
}
\arguments{
\item{st_profile}{is the location of the ST_profiles.csv that can be obtained in mlst database}

\item{alleles}{is the data frame generated from alleles}

\item{changeSet}{is the location of the csv created, see \code{changeCurve} for the format of the csv file}

\item{included}{is the ST to be included for analysis}

\item{excluded}{is the ST to be excluded from analysis
Note that included is used before excluded, a range can be used, e.g. to included ST 1 to 100,
included=c(1:100), can be similarly applied to excluded}
}
\value{
data frame with ST, fragments sequences, fragments curve numbers
}
\description{
\code{combineLoci} is used to organise the fragment to its respective ST
}
