% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{distinct}
\alias{distinct}
\title{\code{distinct} is used to calculate the percentage of dissimilarity.}
\usage{
distinct(seq, tNames, exc = NULL)
}
\arguments{
\item{seq}{the pattern generated from \code{percent.pattern}}

\item{tNames}{is the name of the group of interested alleles.}

\item{exc}{is the original excluded positions, additional positions will be added to it.}
}
\value{
Will returns percentage of dissimilarity.
}
\description{
\code{distinct} is used to calculate the percentage of dissimilarity.
}
