% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{flagAllele}
\alias{flagAllele}
\title{\code{flagAllele} is used to find out a list of allelic profiles
that has been flagged and will not be included in computation
of minimum SNPs.}
\usage{
flagAllele(seq)
}
\arguments{
\item{seq}{a list of SeqFastadna. To keep it simple, 
use read.fasta from seqinr to import the fasta file.}
}
\value{
Will return a list of ignored allelic profiles.
}
\description{
\code{flagAllele} is used to find out a list of allelic profiles
that has been flagged and will not be included in computation
of minimum SNPs.
}
\examples{
flagAllele(Chlamydia)
}
