% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{flagPosition}
\alias{flagPosition}
\title{\code{flagPosition} is used to find out positions that 
will be ignored in calculation (either not A,C,G,T or '-')}
\usage{
flagPosition(proSeq, dashIgnore = TRUE)
}
\arguments{
\item{proSeq}{Sequences after processed, i.e. all with the same length}

\item{dashIgnore}{whether to treat '-' as another type}
}
\value{
Will return a list of positions that need to be ignored.
}
\description{
\code{flagPosition} is used to find out positions that 
will be ignored in calculation (either not A,C,G,T or '-')
}
