% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrm.R
\name{generateID}
\alias{generateID}
\title{\code{generateID} is used to generate the MElT ID}
\usage{
generateID(c_st_all)
}
\arguments{
\item{c_st_all}{is the data frame from changeCurve}
}
\value{
the same data frame with MElT ID appended to the last column
}
\description{
\code{generateID} is used to generate the MElT ID
}
