% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrm.R
\name{hrm.dIndex}
\alias{hrm.dIndex}
\title{\code{hrm.dIndex} returns the simpson index of the result based on the MElT-ID}
\usage{
hrm.dIndex(result)
}
\arguments{
\item{result}{is the result from \code{generateID}}
}
\value{
will return simpson index
}
\description{
\code{hrm.dIndex} returns the simpson index of the result based on the MElT-ID
}
