% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrm.R
\name{hrm.pattern}
\alias{hrm.pattern}
\title{\code{hrm.pattern} is used to generate the pattern for calculating simpson index}
\usage{
hrm.pattern(result)
}
\arguments{
\item{result}{the result from \code{generateID}}
}
\value{
a pattern that can be input into \code{simpson.calculate}
}
\description{
\code{hrm.pattern} is used to generate the pattern for calculating simpson index
}
