% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrm.R
\name{invalidSeq}
\alias{invalidSeq}
\title{\code{invalidSeq} is used to check if the sequence contains invalid character}
\usage{
invalidSeq(proSeq, acceptedChar = c("A", "a", "C", "c", "T", "t", "G",
  "g", "-"))
}
\arguments{
\item{proSeq}{is the sequence to be calculated}

\item{acceptedChar}{is the characters which are considered valid}
}
\value{
will returns the sum of GC content in the sequence
}
\description{
\code{invalidSeq} is used to check if the sequence contains invalid character
}
