% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{output.percent}
\alias{output.percent}
\title{\code{output.percent} is used to output the result to csv file.}
\usage{
output.percent(result, target, seq, filename = NULL)
}
\arguments{
\item{result}{is from \code{branch.percent}}

\item{target}{is the name of the allele targeted}

\item{seq}{is fastaDNA object to analysed.}

\item{filename}{is the name of the output csv file, default to datetime of when the file is generated}
}
\value{
will generate a csv file with the filename in the current directory
}
\description{
\code{output.percent} is used to output the result to csv file.
}
