% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMode.R
\name{output.simpson}
\alias{output.simpson}
\title{\code{output.simpson} is used to output the result to csv file.}
\usage{
output.simpson(result, seq, filename = NULL)
}
\arguments{
\item{result}{is from \code{branch.simpson}}

\item{seq}{is fastaDNA object to analysed.}

\item{filename}{is the name of the output csv file, default to datetime of when the file is generated}
}
\value{
will generate a csv file with the filename in the current directory
}
\description{
\code{output.simpson} is used to output the result to csv file.
}
