% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{percent.calculate}
\alias{percent.calculate}
\title{\code{percent.calculate} is used to calculate the percentage of dissimilarity.}
\usage{
percent.calculate(pattern, tNames)
}
\arguments{
\item{pattern}{is the structure generated from \code{percent.pattern}.}

\item{tNames}{is the name of the reference allele.}
}
\value{
Will returns percentage of dissimilarity.
}
\description{
\code{percent.calculate} is used to calculate the percentage of dissimilarity.
}
