% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{percent.pattern}
\alias{percent.pattern}
\title{\code{percent.calculate} is used to find the calculate the percentage of 
similarity at alleles.}
\usage{
percent.pattern(seq, step, positions)
}
\arguments{
\item{seq}{the pattern generated from \code{percent.pattern}}

\item{step}{the pattern generated either from previous function invocation or empty list}

\item{tName}{the position to add to the pattern}
}
\value{
Will return the pattern for analysis later
}
\description{
\code{percent.calculate} is used to find the calculate the percentage of 
similarity at alleles.
}
