% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{percent.residual}
\alias{percent.residual}
\title{\code{percent.residual} is used to find the residual.}
\usage{
percent.residual(seq, targets, positions = NULL, excluded = NULL)
}
\arguments{
\item{seq}{is fastaDNA object to analysed.}

\item{targets}{is the name of the reference allele.}

\item{positions}{is the positions from the result.}

\item{excluded}{is the excluded positions}
}
\value{
Will returns the residual positions.
}
\description{
\code{percent.residual} is used to find the residual.
}
