% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMode.R
\name{present.simpson}
\alias{present.simpson}
\title{\code{present.simpson} is used to present the result from the calculation of simpson's index.}
\usage{
present.simpson(seq, result)
}
\arguments{
\item{seq}{is fastaDNA object to analysed.}

\item{result}{is the result from \code{branch.simpson}.}
}
\value{
Will returns the presentation of the result.
}
\description{
\code{present.simpson} is used to present the result from the calculation of simpson's index.
}
