% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{processAllele}
\alias{processAllele}
\title{\code{processAllele} is used to returned the processed allelic
profiles.}
\usage{
processAllele(seq)
}
\arguments{
\item{seq}{a list of SeqFastadna. To keep it simple, 
use read.fasta from seqinr to import the fasta file.}
}
\value{
Will return the processed allelic profiles.
}
\description{
\code{processAllele} is used to returned the processed allelic
profiles.
}
