% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{read.fasta}
\alias{read.fasta}
\title{\code{read.fasta} override the default read.fasta provided seqinr
The parameters are exactly the same as the parameters from the original}
\usage{
read.fasta(file, seqtype = c("DNA", "AA"), as.string = FALSE,
  forceDNAtolower = TRUE, set.attributes = TRUE, legacy.mode = TRUE,
  seqonly = FALSE, strip.desc = FALSE, bfa = FALSE,
  sizeof.longlong = .Machine$sizeof.longlong,
  endian = .Platform$endian, apply.mask = TRUE)
}
\description{
\code{read.fasta} override the default read.fasta provided seqinr
The parameters are exactly the same as the parameters from the original
}
