% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{readFasta}
\alias{readFasta}
\title{\code{readFasta} is used to read fasta file}
\usage{
readFasta(file)
}
\arguments{
\item{file}{file path}
}
\value{
Will return fastaDNA.
}
\description{
\code{readFasta} is used to read fasta file
}
