% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrm.R
\name{setID}
\alias{setID}
\title{\code{setID} reads from csv file and out}
\usage{
setID(newLocation, result = NULL)
}
\arguments{
\item{newLocation}{is the location of the csv file}

\item{result}{is the result from \code{generateID}}
}
\value{
will return the new dataframe with MElT-ID appended
}
\description{
\code{setID} reads from csv file and out
}
