% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{similar.percent}
\alias{similar.percent}
\title{\code{similar.percent} is used to find the calculate the percentage of 
similarity at alleles.}
\usage{
similar.percent(seq, ref)
}
\arguments{
\item{seq}{a list of SeqFastadna. To keep it simple, 
use read.fasta from seqinr to import the fasta file.}

\item{ref}{the specific allele to be identified.}
}
\value{
Will return the a list of SNPs that can be used 
and the associated percentage at the particular location.
\dontrun{
similar.percent(a, b) where a is an allelic profiles 
which have not been checked.
}
}
\description{
\code{similar.percent} is used to find the calculate the percentage of 
similarity at alleles.
}
\examples{
A_D213=chlamydia[[1]]
similar.percent(chlamydia, 'A_D213')
}
