% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{similar}
\alias{similar}
\title{\code{similar} is used to find the 100 SNPs 
that can be used to differentiate the allele from other alleles
and also the percentage associated with the particular SNP.}
\usage{
similar(seq)
}
\arguments{
\item{seq}{a list of SeqFastadna. To keep it simple, 
use read.fasta from seqinr to import the fasta file.}
}
\value{
Will return the a list of SNPs (100) that can be used and the associated percentage.
}
\description{
\code{similar} is used to find the 100 SNPs 
that can be used to differentiate the allele from other alleles
and also the percentage associated with the particular SNP.
}
\examples{
similar(chlamydia)
}
