% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMode.R
\name{similar.simpson}
\alias{similar.simpson}
\title{\code{similar.simpson} is used to calculate the simpson index and list the position with highest index.}
\usage{
similar.simpson(seq, level = 1, included = NULL, excluded = NULL)
}
\arguments{
\item{seq}{is fastaDNA object to analysed.}

\item{level}{is the number of positions.}

\item{included}{is the included position for analysed, these will force the computation
to compute the simpson's index at the position no matter what}

\item{excluded}{is the positions that are excluded from computation}
}
\value{
Will returns simpson's index and the position.
}
\description{
\code{similar.simpson} is used to calculate the simpson index and list the position with highest index.
}
