% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMode.R
\name{simpson.calculate}
\alias{simpson.calculate}
\title{\code{simpson.calculate} is used to calculate the simpson's index given a pattern.}
\usage{
simpson.calculate(pattern, N)
}
\arguments{
\item{pattern}{is a pattern, can be a vector or a list.}

\item{N}{is the total number of entities that are in the pattern.}
}
\value{
Will returns the simpson's index for the pattern.
}
\description{
\code{simpson.calculate} is used to calculate the simpson's index given a pattern.
}
