% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMode.R
\name{simpson.pattern}
\alias{simpson.pattern}
\title{\code{simpson.pattern} is used to generate pattern for calculation at a later stage.}
\usage{
simpson.pattern(seq, position, appended = NULL)
}
\arguments{
\item{seq}{is fastaDNA object to analysed.}

\item{position}{is the position of the sequences used to generate pattern.}

\item{appended}{is the pattern that the current operation will appended onto}
}
\value{
Will returns the generated pattern.
}
\description{
\code{simpson.pattern} is used to generate pattern for calculation at a later stage.
}
