% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{usualLength}
\alias{usualLength}
\title{\code{usualLength} is used to find out 
the length of the sequence (W/O deletion).}
\usage{
usualLength(seq)
}
\arguments{
\item{seq}{a list of SeqFastadna. To keep it simple, 
use read.fasta from seqinr to import the fasta file.}
}
\value{
Will return the maximum length of all the allelic profiles.
}
\description{
\code{usualLength} is used to find out 
the length of the sequence (W/O deletion).
}
\examples{
sample.case1 <-function(){
	#Read the file
	Chlamydia <- read.fasta(file='../data/Chlamydia_mapped.txt')
	#Process the file
	Chlamydia2 <- processAllele(Chlamydia)
	untempered <-read.fasta(file='../data/Chlamydia_mapped.txt')
	
	#Since Chlamydia is normal
	checkIdentical(Chlamydia2, untempered)
}

sample.case2 <-function(){
}
}
