% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{Chlamydia_2}
\alias{Chlamydia_2}
\title{Normal Fasta sample file
Chlamydia sequence type with multiple deletions}
\description{
Normal Fasta sample file
Chlamydia sequence type with multiple deletions
}
