% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{present.percent}
\alias{present.percent}
\title{\code{present.percent} is used to present the result of percent calculation.}
\usage{
present.percent(seq, target, result)
}
\arguments{
\item{seq}{is fastaDNA object to analysed.}

\item{target}{is the name of the allele targeted}

\item{result}{the result from \code{branch.percent}}
}
\value{
Will print out the result
}
\description{
\code{present.percent} is used to present the result of percent calculation.
}
