% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{present.percent}
\alias{present.percent}
\title{\code{}
\code{present.percent} is used to find present and filter the 
similarity calculated using \code{similar.percent}.}
\usage{
present.percent(result, percent = 100, number = 100)
}
\arguments{
\item{result}{the result from \code{similar.percent}.}

\item{percent}{minimum percentage to be included}

\item{number}{number of results to be displayed}
}
\value{
Will return the a list of SNPs (as specified) that can be used 
and the associated percentage at the particular location.
}
\description{
\code{}
\code{present.percent} is used to find present and filter the 
similarity calculated using \code{similar.percent}.
}
