% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{similar.percent}
\alias{similar.percent}
\title{\code{similar.percent} is used to calculate the percentage of dissimilarity and list the position with highest percentage.}
\usage{
similar.percent(seq, target, level = 1, included = NULL, excluded = NULL)
}
\arguments{
\item{seq}{is fastaDNA object to analysed.}

\item{target}{is the name of the reference allele}

\item{level}{is the number of positions.}

\item{included}{is the included position for analysed, these will force the computation
to compute the percentage of dissimilarity at the position no matter what}

\item{excluded}{is the positions that are excluded from computation}
}
\value{
Will returns simpson's index and the position.
}
\description{
\code{similar.percent} is used to calculate the percentage of dissimilarity and list the position with highest percentage.
}
