% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMode.R
\name{simpson.calculate}
\alias{simpson.calculate}
\title{\code{similar.simpson} is used to find the calculate the percentage of 
similarity at alleles.}
\usage{
simpson.calculate(pattern, N)
}
\arguments{
\item{ref}{the specific allele to be identified.}
}
\value{
Will return the a list of SNPs that can be used 
and the associated percentage at the particular location.
}
\description{
\code{similar.simpson} is used to find the calculate the percentage of 
similarity at alleles.
}
\examples{
similar.simpson(chlamydia)
}
