% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{usualLength}
\alias{usualLength}
\title{\code{usualLength} is used to find out 
the length of the sequence (W/O deletion).}
\usage{
usualLength(seq)
}
\arguments{
\item{seq}{a list of SeqFastadna. To keep it simple, 
use read.fasta from seqinr to import the fasta file.}
}
\value{
Will return the maximum length of all the allelic profiles.
}
\description{
\code{usualLength} is used to find out 
the length of the sequence (W/O deletion).
}
\examples{
library(seqinr)
chlamydia <- read.fasta('Chlamydia.fas')
usualLength(chlamydia)
}
