% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{usualLength}
\alias{usualLength}
\title{\code{usualLength} is used to find out 
the length of the sequence (W/O deletion).}
\usage{
usualLength(seq)
}
\arguments{
\item{seq}{a list of SeqFastadna. To keep it simple, 
use read.fasta from seqinr to import the fasta file.}
}
\value{
Will return the maximum length of all the allelic profiles.
}
\description{
\code{usualLength} is used to find out 
the length of the sequence (W/O deletion).
}
\examples{
sample.case1 <-function(){
	#Read the file
	Chlamydia <- read.fasta(file='../data/Chlamydia_mapped.txt')
	
	#STEP 1. Process the file
	Chlamydia <- processAllele(Chlamydia)
	untempered <-read.fasta(file='../data/Chlamydia_mapped.txt')
	
	#Since Chlamydia is normal
	checkIdentical(Chlamydia, untempered)
	
	#PERCENT MODE
	result=similar.percent(Chlamydia, 'A_D213')
	present=present.percent(result, 98, 100)
	
	#All result should have percent higher or equal to 98
	for (a in 1:100){
	checkTrue(present[[a]]$percent>=98)
	}
	
	#SIMPSON MODE
	result=branch.simpson(Chlamydia, level=1, numRes=3)
	output=present.simpson(Chlamydia, result)
	
	#Should have 3 results
	checkTrue(length(output)==3)
	
	checkEquals(output[[1]]$'Index',  0.7344, tolerance=0.00016)
	Description=paste('At position:', '1988', sep='-')
	checkEquals(output[[1]]$'Description', Description)
	
	checkEquals(output[[2]]$'Index', 0.7318, tolerance=0.00016)
	Description=paste('At position:', '2044', sep='-')
	checkEquals(output[[2]]$'Description', Description)
	
	checkEquals(output[[3]]$'Index', 0.7266, tolerance=0.00016)
	Description=paste('At position:', '2034', sep='-')
	checkEquals(output[[3]]$'Description', Description)
}

sample.case2 <-function(){

}

test.deactivation <- function()
{
  DEACTIVATED('Deactivating integration test function')
}
}
